+++ NEWS.md
@@ -1,8 +1,9 @@
 # CHANGES TO margins 0.3.0 #
 
-## margins 0.2.14
+## margins 0.2.15
 
-* Tentatively exposed previously internal `mfx_*()` functions to provide variable-specific marginal effects calculations. (#31)
+* Expose previously internal `mfx()` generic and methods to provide variable-specific marginal effects calculations. (#31)
+* Added example dataset from **marfx** package. (#31)
 
 ## margins 0.2.13